src/visual_cues.py
```python
import tkinter as tk

class VisualCues:
    def __init__(self, root):
        self.root = root
        self.canvas = tk.Canvas(self.root, width=800, height=600)
        self.canvas.pack()

    def draw_agent_activity(self, agent_activity):
        for activity in agent_activity:
            self.canvas.create_rectangle(activity['x'], activity['y'], activity['x']+20, activity['y']+20, fill=activity['color'])

    def draw_project_status(self, project_status):
        for status in project_status:
            self.canvas.create_text(status['x'], status['y'], text=status['status'], fill=status['color'])

    def update_visual_cues(self, agent_activity, project_status):
        self.canvas.delete('all')
        self.draw_agent_activity(agent_activity)
        self.draw_project_status(project_status)

if __name__ == "__main__":
    root = tk.Tk()
    visual_cues = VisualCues(root)
    root.mainloop()
```